FEAT : Redis 변경 관련 서비스 적용

+++ concert-reservation-api/src/main/java/io/hhplus/concert/interfaces/dto/QueueStatusResponse.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class QueueStatusResponse {
 
     private String userId;
-    private int position;
+    private Long position;
 
 }
